YUK-7152  STARS - Create Hardware Page Allows Devices to be Created with Serial Numbers Longer than Nine Digits Add range, Update range, Delete range, Configure range and ShipmentSNRange Controllers, Tasks, Jsps updated to use 19-digit long values instead of 10-digit int values.  SelectAll (using *,asterisk) type of functionality removed from Update, Delete Ranges. Validation, error messages made consistent.

+++ yukon-client/stars/src/com/cannontech/stars/core/dao/StarsSearchDao.java
@@ -115,8 +115,8 @@ public interface StarsSearchDao {
 	 * @return List of hardware found
 	 */
 	public List<LiteStarsLMHardware> searchLMHardwareBySerialNumberRange(
-			int startSerialNumber, 
-			int endSerialNumber, 
+			long startSerialNumber, 
+			long endSerialNumber, 
 			int deviceTypeDefinitionId, 
 			List<LiteStarsEnergyCompany> energyCompanyList);
